requirements.txt
streamlit==1.32.0
langchain==0.1.20
langchain-community==0.0.38
PyMuPDF==1.24.0
python-docx==1.1.0
sentence-transformers==2.7.0
faiss-cpu==1.7.4
groq==0.9.0
python-dotenv==1.0.1
numpy==1.26.4